chore: refactor for OptimizeDiskPerformance code
move dir

fix golint

fix echoToFile
